--- service/order/src/main/java/jabaclass/order/common/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package jabaclass.order.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
