--- service/order/src/main/java/jabaclass/order/common/error/BusinessException.java
@@ -1,22 +1,0 @@
-package jabaclass.order.common.error;
-
-import org.springframework.http.HttpStatus;
-
-public class BusinessException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public BusinessException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public HttpStatus getStatus() {
-        return errorCode.getStatus();
-    }
-
-    @Override
-    public String getMessage() {
-        return errorCode.getMessage();
-    }
-}
