--- service/order/src/main/java/jabaclass/order/common/error/CommonErrorCode.java
@@ -1,27 +1,0 @@
-package jabaclass.order.common.error;
-
-import org.springframework.http.HttpStatus;
-
-public enum CommonErrorCode implements ErrorCode {
-
-    INVALID_PARAMETER(HttpStatus.BAD_REQUEST, "파라미터 값을 확인해주세요."),
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버 에러입니다. 서버 팀에 연락주세요!");
-
-    private final HttpStatus status;
-    private final String message;
-
-    CommonErrorCode(HttpStatus status, String message) {
-        this.status = status;
-        this.message = message;
-    }
-
-    @Override
-    public HttpStatus getStatus() {
-        return status;
-    }
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-}
