--- service/order/src/main/java/jabaclass/order/common/error/ErrorCode.java
@@ -1,10 +1,0 @@
-package jabaclass.order.common.error;
-
-import org.springframework.http.HttpStatus;
-
-public interface ErrorCode {
-
-    HttpStatus getStatus();
-
-    String getMessage();
-}
