+++ service/order/src/main/java/jabaclass/order/domain/model/OrderStatus.java
@@ -1,0 +1,9 @@
+package jabaclass.order.domain.model;
+
+public enum OrderStatus {
+    PENDING,
+    PAID,
+    CANCELED,
+	FAILED,
+    REFUNDED
+}
