+++ service/order/src/main/java/jabaclass/order/domain/repository/OrderRepository.java
@@ -1,0 +1,13 @@
+package jabaclass.order.domain.repository;
+
+import jabaclass.order.domain.model.Order;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface OrderRepository {
+
+    Order save(Order order);
+
+    Optional<Order> findById(UUID orderId);
+}
