+++ service/order/src/main/java/jabaclass/order/infrastructure/persistence/OrderJpaRepository.java
@@ -1,0 +1,9 @@
+package jabaclass.order.infrastructure.persistence;
+
+import jabaclass.order.domain.model.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface OrderJpaRepository extends JpaRepository<Order, UUID> {
+}
