+++ service/order/src/main/java/jabaclass/order/infrastructure/persistence/OrderRepositoryAdapter.java
@@ -1,0 +1,26 @@
+package jabaclass.order.infrastructure.persistence;
+
+import jabaclass.order.domain.model.Order;
+import jabaclass.order.domain.repository.OrderRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+@RequiredArgsConstructor
+public class OrderRepositoryAdapter implements OrderRepository {
+
+    private final OrderJpaRepository orderJpaRepository;
+
+    @Override
+    public Order save(Order order) {
+        return orderJpaRepository.save(order);
+    }
+
+    @Override
+    public Optional<Order> findById(UUID orderId) {
+        return orderJpaRepository.findById(orderId);
+    }
+}
