--- service/order/src/main/java/jabaclass/order/order/domain/model/OrderStatus.java
@@ -1,9 +1,0 @@
-package jabaclass.order.order.domain.model;
-
-public enum OrderStatus {
-    PENDING,
-    PAID,
-    CANCELED,
-	FAILED,
-    REFUNDED
-}
