--- service/order/src/main/java/jabaclass/order/order/infrastructure/persistence/OrderRepositoryAdapter.java
@@ -1,38 +1,0 @@
-package jabaclass.order.order.infrastructure.persistence;
-
-import jabaclass.order.order.domain.model.Order;
-import jabaclass.order.order.domain.model.OrderStatus;
-import jabaclass.order.order.domain.repository.OrderRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-@RequiredArgsConstructor
-public class OrderRepositoryAdapter implements OrderRepository {
-
-    private final OrderJpaRepository orderJpaRepository;
-
-    @Override
-    public Order save(Order order) {
-        return orderJpaRepository.save(order);
-    }
-
-    @Override
-    public Optional<Order> findById(UUID orderId) {
-        return orderJpaRepository.findById(orderId);
-    }
-
-    @Override
-    public List<Order> findAllByUserId(UUID userId) {
-        return orderJpaRepository.findAllByUserId(userId);
-    }
-
-    @Override
-    public List<Order> findAllByUserIdAndStatus(UUID userId, OrderStatus status) {
-        return orderJpaRepository.findAllByUserIdAndStatus(userId, status);
-    }
-}
