--- service/order/src/main/java/jabaclass/order/order/presentation/dto/request/CancelOrderRequestDto.java
@@ -1,11 +1,0 @@
-package jabaclass.order.order.presentation.dto.request;
-
-import java.util.UUID;
-
-import jakarta.validation.constraints.NotNull;
-
-public record CancelOrderRequestDto(
-    @NotNull
-    UUID userId
-) {
-}
